Trial increments is now set in configuration file and outputed + new scenes for timing diagnostics
There is now the option to set "TrialInitialValue" to an integer from which the timing system will use to invert values. The trial counter will also be outputted in the csv after each trial. TODO: currently, only think outputted is the number of the trial when a 3D trial is happening, this should change and output for every trial.